Add createCard function to MovieList.js
