+++ src/components/MovieList.js
@@ -1,5 +1,23 @@
-import createCard from './Card.js'; // Import createCard from Card.js
+// import createCard from './Card.js'; // Import createCard from Card.js
 import InvolvementAPI from './InvolvementAPI.js';
+// Create cards
+const createCard = (show, initialLikes = 0) => {
+  const card = document.createElement('div');
+  card.classList.add('card');
+  card.innerHTML = `
+        <img src="${show.image.medium}" alt="${show.name}">
+        <div class="card-content">
+            <h3>${show.name}</h3>
+        </div>
+        <div class="card-buttons">
+            <button class="like-button" data-show-id="${show.id}" data-likes="${initialLikes}">
+                ❤️ Like (${initialLikes})
+            </button>
+            <button class="comment-button">💬 Comment</button>
+        </div>
+    `;
+  return card;
+};
 
 const API_URL = 'https://api.tvmaze.com/shows';
